pyproject.toml
[project]
name = "mavspoi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "openai>=1.40.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
]
